refactor: 트리 구조 개선
