+++ src/main/java/com/ssafy/jjtrip/domain/triplog/controller/TripLogController.java
@@ -4,13 +4,13 @@ import com.ssafy.jjtrip.common.dto.PageDto;
 import com.ssafy.jjtrip.common.dto.SliceDto;
 import com.ssafy.jjtrip.common.s3.dto.PresignedUrlResponseDto;
 import com.ssafy.jjtrip.common.security.CustomUserDetails;
-import com.ssafy.jjtrip.domain.triplog.dto.ImageUploadRequestDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogCommentRequestDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogCreateRequestDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogCreateResponseDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogFeedResponseDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogLikeResponseDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogUpdateRequestDto;
+import com.ssafy.jjtrip.domain.triplog.dto.request.ImageUploadRequestDto;
+import com.ssafy.jjtrip.domain.triplog.dto.request.TripLogCommentRequestDto;
+import com.ssafy.jjtrip.domain.triplog.dto.request.TripLogCreateRequestDto;
+import com.ssafy.jjtrip.domain.triplog.dto.request.TripLogUpdateRequestDto;
+import com.ssafy.jjtrip.domain.triplog.dto.response.TripLogCreateResponseDto;
+import com.ssafy.jjtrip.domain.triplog.dto.response.TripLogFeedResponseDto;
+import com.ssafy.jjtrip.domain.triplog.dto.response.TripLogLikeResponseDto;
 import com.ssafy.jjtrip.domain.triplog.service.TripLogCommentService;
 import com.ssafy.jjtrip.domain.triplog.service.TripLogImageService;
 import com.ssafy.jjtrip.domain.triplog.service.TripLogLikeService;
