+++ src/main/java/com/ssafy/jjtrip/domain/triplog/dto/request/ImageUploadRequestDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.jjtrip.domain.triplog.dto;
+package com.ssafy.jjtrip.domain.triplog.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
