+++ src/main/java/com/ssafy/jjtrip/domain/triplog/dto/request/TripLogCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.jjtrip.domain.triplog.dto;
+package com.ssafy.jjtrip.domain.triplog.dto.request;
 
 import com.ssafy.jjtrip.domain.triplog.entity.TripLogVisibility;
 import jakarta.validation.constraints.NotBlank;
