+++ src/main/java/com/ssafy/jjtrip/domain/triplog/dto/response/TripLogCommentResponseDto.java
@@ -1,7 +1,6 @@
-package com.ssafy.jjtrip.domain.triplog.dto;
+package com.ssafy.jjtrip.domain.triplog.dto.response;
 
 import java.time.LocalDateTime;
-
 import java.util.List;
 
 public record TripLogCommentResponseDto(
