+++ src/main/java/com/ssafy/jjtrip/domain/triplog/dto/response/TripLogCreateResponseDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.jjtrip.domain.triplog.dto;
+package com.ssafy.jjtrip.domain.triplog.dto.response;
 
 public record TripLogCreateResponseDto(
         Long logId
