+++ src/main/java/com/ssafy/jjtrip/domain/triplog/dto/response/TripLogLikeResponseDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.jjtrip.domain.triplog.dto;
+package com.ssafy.jjtrip.domain.triplog.dto.response;
 
 public record TripLogLikeResponseDto(
     boolean liked,
