+++ src/main/java/com/ssafy/jjtrip/domain/triplog/dto/response/TripLogSummaryDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.jjtrip.domain.triplog.dto;
+package com.ssafy.jjtrip.domain.triplog.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
