+++ src/main/java/com/ssafy/jjtrip/domain/triplog/service/TripLogService.java
@@ -3,13 +3,13 @@ package com.ssafy.jjtrip.domain.triplog.service;
 import com.ssafy.jjtrip.common.dto.PageDto;
 import com.ssafy.jjtrip.common.dto.SliceDto;
 import com.ssafy.jjtrip.domain.trip.service.TripService;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogCommentResponseDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogCreateRequestDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogCreateResponseDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogDetailResponseDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogFeedResponseDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogImageResponseDto;
-import com.ssafy.jjtrip.domain.triplog.dto.TripLogUpdateRequestDto;
+import com.ssafy.jjtrip.domain.triplog.dto.request.TripLogCreateRequestDto;
+import com.ssafy.jjtrip.domain.triplog.dto.request.TripLogUpdateRequestDto;
+import com.ssafy.jjtrip.domain.triplog.dto.response.TripLogCommentResponseDto;
+import com.ssafy.jjtrip.domain.triplog.dto.response.TripLogCreateResponseDto;
+import com.ssafy.jjtrip.domain.triplog.dto.response.TripLogDetailResponseDto;
+import com.ssafy.jjtrip.domain.triplog.dto.response.TripLogFeedResponseDto;
+import com.ssafy.jjtrip.domain.triplog.dto.response.TripLogImageResponseDto;
 import com.ssafy.jjtrip.domain.triplog.entity.TripLog;
 import com.ssafy.jjtrip.domain.triplog.entity.TripLogVisibility;
 import com.ssafy.jjtrip.domain.triplog.exception.TripLogErrorCode;
